+++ bindings/web/types/lz-string.d.ts
@@ -1,0 +1,3 @@
+declare module "lz-string/libs/lz-string.js" {
+    function decompressFromUTF16(input: string): string;
+}