+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `--capability` command-line argument (used as `rune run ./whatever.rune
   --capability image:person.png`) has been removed after emitting a warning for
   several versions ([#256](https://github.com/hotg-ai/rune/issues/256))
+- The old `Runefile` format (the text-based domain-specific language, not the
+  YAML version) has now been removed
 
 ## [0.5.3] - 2021-08-11